Rename `FirstSteps` -> `StepVerifiers`
Motivation:

`FirstSteps` is harder to discover compare to `StepVerifiers` factory.
Follow up for #1240.
